Update keyword list from spec
This uses the spec at http://www.contrib.andrew.cmu.edu/~shadow/sql/sql1992.txt
for the keyword list, and also expands the keyword exceptions in order for the
SQL to come out nicely. This change requires a list of keyword exceptions, to a
method to provide that was added, and every gutter check call had to be updated
for that.