This deals with the IF keyword
Interestingly, the IF keyword isn't part of the original keyword list, I've
added the 'possibly reserved in the future' to my list of leywords, and also
added 'replace' and 'substring' to the list of keyword exceptions, since those
are function calls.

The output of a tidied IF statement is not optimal, but, like the CREATE TABLE
statement, it may have to have a different formatting strategy.